Fix mis-handling of Condition
`Condition` is not an intrinsic function in the same way as the others;
it does not need special handling as the start of a mapping, nor to be
translated to/from `!Condition`. In fact, as far as the encoding and
flip is concerned, it's just another opaque key-value.

This bug only hasn't impacted us yet because all our current use-cases
have the `Condition` element coming out not-first in its enclosing
resource `MappingStart`; and this prevents it from being (mis-)handled
as the (non-existent) intrinsic `!Condition`.

By pure happenstance, the update to aeson-2.0 that just occurred in a
downstream application caused re-ordering of Object keys and placed one
of our the `Condition` keys first in the resource, causing us to hit
this bug. What's more, the result of the bug is a completely blank
string coming out the other side!